+++ R/rd--plot.R
@@ -3,6 +3,7 @@ plot_rd <- function(simulation,
                     method = c("paf", "rq"),
                     element_id = NULL,
                     dygraph_group = NULL,
+                    chemical_ids = NULL,
                     ...)
 {
   assert_erahumed_simulation(simulation)
@@ -26,6 +27,7 @@ plot_rd <- function(simulation,
             type = type,
             method = method,
             dygraph_group = dygraph_group,
-            chemical_db = get_etc(simulation, "chemical_db")
+            chemical_db = get_etc(simulation, "chemical_db"),
+            chemical_ids = chemical_ids
             )
 }