+++ R/rl--plot.R
@@ -2,6 +2,7 @@ plot_rl <- function(simulation,
                     type = c("chronic", "acute"),
                     method = c("paf", "rq"),
                     dygraph_group = NULL,
+                    chemical_ids = NULL,
                     ...)
 {
   assert_erahumed_simulation(simulation)
@@ -14,6 +15,7 @@ plot_rl <- function(simulation,
             type = type,
             method = method,
             dygraph_group = dygraph_group,
-            chemical_db = get_etc(simulation, "chemical_db")
+            chemical_db = get_etc(simulation, "chemical_db"),
+            chemical_ids = chemical_ids
             )
 }